requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==37.0.4
Deprecated==1.2.13
Django==3.2.14
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dnspython==2.2.1
eventlet==0.33.1
greenlet==1.1.2
json-tools==0.4.1
kombu==5.2.4
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.23.1
opencv-python==4.4.0.46
packaging==21.3
pandas==1.4.3
prompt-toolkit==3.0.30
pycparser==2.21
pycryptodome @ file:///C:/ci/pycryptodome_1658328398039/work
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wincertstore==0.2
wrapt==1.14.1
